requirements.txt
streamlit
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python-headless
mediapipe
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn
tqdm
pandas
timm>=0.9.0
efficientnet-pytorch
mlflow
kaggle
python-multipart
scipy>=1.10.0
albumentations>=1.3.0
datasets>=2.14.0
huggingface-hub>=0.17.0